Right Side Auton
This is an unfinished, rough shell of what our right side autonomous will be. In the future, we will collect more data on what we want this to be, and the code that. I have not had a large conversation with the group about what we want to do, but I know that we probably want to score a pre-loaded specimen on the low bar, and score some samples in the observation zone.