79 Error Fix and FCD Motor Power
Fixed the 79 errors that appeared in claw  by reverting it to an older version I found on github. If anybody added that extra code for a reason, I am sorry, but it added to many problems. I also tuned the motor powers on the robot, so that the 4-bar on clippy may actually lift itself, and the flywheel will be more powerful.